+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - `AgroSuccessFuelMoistureRiskTable` to convert vegetationMoistureParam to
   fuelMoistureRiskFactor
+- `GridPointIterable` class added to `RepastGridUtils`
 
 #### CHANGED